requirements.txt
# RTACE - Real-Time Transaction Anomaly & Containment Engine
# Python 3.11

kafka-python==2.0.2
redis==5.0.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
prometheus_client==0.19.0
